Deleted a Notes File and Added The code in Notes File to Homejs File and to make a condition and added to Routes.js file
