+++ src/App.js
@@ -1,11 +1,12 @@
-import 'bootstrap/dist/css/bootstrap.min.css';
-import './App.css';
-import Home from './components/Home';
+import "bootstrap/dist/css/bootstrap.min.css";
+import "./App.css";
+import Home from "./components/Home";
+import RouterBages from "./Routes/Router";
 
 function App() {
   return (
     <div className="App">
-      <Home/>
+      <RouterBages />
     </div>
   );
 }
